move shared session type into its own module
