+++ web/src/lib/session-types.ts
@@ -1,0 +1,6 @@
+export interface StoredSession {
+  sessionToken: string
+  keyId: string
+  displayName: string
+  expiresAt: number
+}
